+++ eslint.config.cjs
@@ -1,0 +1,32 @@
+const { FlatCompat } = require('@eslint/eslintrc');
+const js = require('@eslint/js');
+
+const compat = new FlatCompat({
+    baseDirectory: __dirname,
+    recommendedConfig: js.configs.recommended,
+    allConfig: js.configs.all,
+});
+
+module.exports = [
+    ...compat.config({
+        env: {
+            es2021: true,
+            node: true,
+            mocha: true,
+        },
+        parser: '@typescript-eslint/parser',
+        parserOptions: {
+            ecmaVersion: 'latest',
+            sourceType: 'module',
+        },
+        plugins: ['@typescript-eslint'],
+        extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
+        rules: {
+            '@typescript-eslint/no-unused-vars': 'off',
+            '@typescript-eslint/no-explicit-any': 'off',
+            '@typescript-eslint/no-require-imports': 'off',
+            'prefer-const': 'off',
+            'no-useless-escape': 'off',
+        },
+    }),
+];